incorporate amazonEcrLogin
